Changed test threshold back to a value compatible with commons math 3.0.
The value had been changed when an improvement in Apache Commons Math
was included in revision 1346513 as part of solving issue MATH-801. The
error dropped from 1.2e-7 to 2.6e-8 so the test was adapted. However,
the fix will be released only as part of Apache Commons Math 3.1 and
Orekit depend on Apache Commons Math 3.0, so the test did not pass
anymore with the official version.