Temporarily added an HermiteInterpolator class.
Note that this class is not expected to remain in Orekit. It
rather belongs to Apache Commons Math. However, since as of writing
Orekit depends on version 3.0 of Apache Commons Math and this class is
not provided by this version of the library, the class is implemented as
a package private class in Orekit. Once the class is included in an
officially released version of Apache Commons Math, it will be removed
from Orekit.