Introduced TransformProvider into the Frame class.
Instead of having a mutable Transform field inside the Frame class and
set/update methods to reset this field, we now use an intermediate
TransformProvider that computes the transform on the fly.

This simplifies the design by separating the fixed topology
(parent/child) from the time-dependent geometry. The topology notion
remains in the Frame class itself, whereas the time-dependent geometry
notion goes into several simpler and specific implementations of the
TransformProvider class.

Despite being a large change affecting a lot of classes, it should not
be visible to most users. The affected classes correspond to private
Orekit classes which represent implementation details and are not part
of the public API. The only public change is that the setTransform and
updateFrame methods are not in the base Frame class anymore, but are in
a new UpdatableFrame class. The setTransform method was mainly
intened to be called by Orekit itself as part of implementing
getTransformTo but is not used anymore. From a caller point of view, in
the very rare cases where these methods were directly called for users
frames, the user code can be adapted by changing the user frames base
classes from Frame to UpdatableFrame. From a callee point of view, these
methods are not usefule anymore and changing the transform from a parent
frame to to child frame should be done by implementing a
TransformProvider rather than implementing an updateFrame method that
calls setTransform. Basically, this means that existing updateFrame
methods must be renamed and return the transform instead of calling
setTransform with it.

This will later (soon) allow to factor all interpolation features from
all time-dependent frames into a single class that will also take care
of multi-threading issues.

Work In Progress concerning issue #3.