Added a thread-safe interpolating transform provider.
This class is intended to be the major component for both thread-safe
*and* efficient frames handling. Simply using "synchronized" statements
does not work for frames as it completely breaks all caching features,
as explained at length in various places in the mailing lists. This
class does not have this drawbacks, it preserve the best of both worlds:
thread safety, cache, and interpolation.