Removed earliest/latest boundaries from TimeStampedGenerator.
Some generators were not able to provide these limits beforehand, as
they discovered their range while generating the data (or as reading it
from a file). The boundary checks are now done a posteriori instead of a
priori.

This change also allowed to set up an exception generators implementors
can use if they cannot generate data for a specific date.

It is now much simpler to implement a TimeStampedGenerator (only one
simple method), and it is possible to do some lazy evaluations if
desired.