Fixed an interpolation problem at boundaries.
When a TimeStampedCache is asked to provide neighbors for a date that is
close to the generator limits (for example near the end of an EOP file),
it triggered an error as soon as the interval was not balanced instead
of having a gradual fallback and still using the available data. This is
especially cumbersome for data that is completely valid until the end of
its interpolation (and sometimes even past the limit for simple linear
models).