Use 4 points interpolation for EOP data following IERS recommendation.
Previously, Orekit used only two points for interpolation.
The expected differences for Earth relative positions with respect to
previous versions are a few millimeters.

Note that now changing the order of the interpolation is a trivial
change as it is defined by a single integer constant
INTERPOLATION_POINTS in the class AbstractEOPHistory.