Replaced local caches by using TransformProvider.
This big change is a great step towards solving issue #3 on
thread-safety issues. All cachedDate/cachedTransform local caches used
by various frames have been removed. They have been replaced by
TransformProvider implementations. For those frames for which the
transform provider is very simple and involves very few computations
(say one sidereal time), a simple dedicated TransfomrProvider
implementation is used. It is far simpler than the original frame code
since it does not cache anything. For the other frames which involves
costrly computations (say precession/nutation models), there is a low
level TransformProvider implementation which does the direct (costly)
computation but does neither cache nor interpolate anything. On top of
that, an InterpolatedTransformProvider adds interpolation and caching
for efficiency, while still ensuring thread-safety.

Note that the interpolation feature does compute angular velocity even
if the underlying raw transform provider does not compute it (it is a
by-product of the polynomial model). This implies that now frames like
TOD or CIRF *do* include the very small precession/nutation velocity
composition effects. This is an improvement but it does change some
results (it is of the order of magnitude of 1.0e-5 to 1.0e-4 m/s in LEO,
higher in GEO). We had to adjust some frames unit test thresholds as
the reference results did not take this effect into account.
as
the reference cases we used