Improved support for INPOP ephemerides files.
The main change introduced here solves the mixed ephemerides
loading identified in issue #75. Now, when a planetary ephemeris
is loaded it will use the solar system barycenter consistent with
the planet (i.e we will not mix anymore JPL and inpop ephemerides if
both are present in the data path).

Other changed introduced here should allow more straightforward use of
INPOP files. The inpop file names are now recognized by default in
addition to the JPL DE file names. The TCB and TDB time scales are
automaticaly handled and ephemerides in astronomical units are also
handled.

This commit solves issue #75 and improves solution for issue #23.