Extended JPL test files to avoid too long tests.
The new thread-safe access to JPL ephemerides tries to bracket the date
between two records. This fails when attempting to access the very last
record in JPL files, so a fallback solution is used, however this take
times as the celestial body always retries to load the missing data.
This problem is really a minor one, as usually people will have JPL
ephemerides that can cover hundreds of years easily, so the attempt to
use bracketing is considered a good thing. Unfortunately, the tests data
was really short and triggers this feature far too often. The tests were
taking significantly longer than usual so we extended the data files to
avoid this.