Split AbstractPropagator abstract class in two layers.
The topmost layer is still called AbstractPropagator and will be shared
amongst both analytical, semi-analytical and numerical propagators. The
new second layer is called AbstractAnalyticalPropagator and is
specialized for analytical propagators (just like the original abstract
class was).

The rational for this change is that we will be able to share some code
with numerical propagators that was not shared up to now. So it is a
preparation for another change that will introduce an
AbstractIntegratedPropagator abstract class on top of
NumericalPropagator first and DSST propagator later on, improving code
sharing between the various classes and hence improving validation and
maintenance.