New package for propagators conversion.
This package provides tools to convert a given propagator or a set of
SpacecraftStates into another propagator.

The conversion principle is to minimize the mean square error for
positions and velocities over a given time span.

The conversion from osculating to mean elements appears as a side
effect of propagation models conversion.

These package extends an original contribution from Telespazio for
TLE (Orbit Converter for Two-Lines Elements) to all propagators.