improved gravity field parsing.
A way to speed up parsing and reduce memory consumption when     
loading gravity fields has been added. Now the user can specify the
maximal degree and order before reading the file. Formerly, the complete
field was parsed and stored on memory, and only afterwards was the array
truncated, when provided to the force model.