Implementation and integration of an immutable TimeStampedCache.
The ImmutableTimeStampedCache will allow fast access to data that is stored in
memory and not modified. It is now used in AbstractEOPHistory, Ephemeris, and
UTCScale. As a side effect the implementation of EOP loaders is now decoupled
from EOPHistory.

The previous implementation of TimeStampedCache has been renamed to
GenericTimeStampedCache, and the TimeStampedCache interface extracted from it.
The getNeighbors(...) method now returns a List instead of an array.