Restrict possibility to ignore EOP.
Ignoring Earth Orientation Parameters corrections generates big errors,
of the order of magnitude of hundreds of meters in low Earth orbit and
kilometers in geostationnary orbit. It has been allowed in Orekit from
the beginning mainly for testing purposes (validation against systems
that did not take EOP into account) and for compatibility with this
legacy systems. Newer systems can't ignore EOP as they often need better
accuracy.

Selecting IERS conventions on the other hand is very fine tuning, of the
order of magnitude of a few millimeters.

So having both choices allowed (with/without EOP and selecting IERS
conventions) at the same time is pointless. If one decides to ignore
EOP, it seems irrelevant to allow selecting the IERS conventions, and it
increases the combinatorial complexity of the frames.

This change mainly separates the choices, depending on the level of
accuracy desired, when retrieving the frames using the FramesFactory
API. If a user concerns are only about crude approximations, then this
user would choose the API with a single boolean parameter applyEOPCorr,
set to false for ignoring the EOP, and set to true for applying the EOP.
With this API, the convention will be hardcoded to IERS 1996, for
compatibility with legacy systems. If a users concerns are about fine
tuning, then this user would choose the API with a single enumerate
parameter conventions, set to IERS_1996, IERS_2003 or IERS_2010
depending on the needs. With this API, the EOP will always be taken into
account, since at this level of accuracy they cannot be ignored.

If we consider for example the GTOD frame, here are the results of the
various factory methods available:

 getGTOD(false)                     --> IERS 1996, EOP ignored
 getGTOD(true)                      --> IERS 1996, EOP considered

 getGTOD(IERSConventions.IERS_1996) --> IERS 1996, EOP considered
 getGTOD(IERSConventions.IERS_2003) --> IERS 2003, EOP considered
 getGTOD(IERSConventions.IERS_2010) --> IERS 2010, EOP considered

Of course in *all* cases, precession and nutation are always considered!

It is by design not possible to get conventions 2003 or 2010 without
considering Earth Orientation Parameters (if you don't care about EOP,
you don't care about differences between modern precession/nutation
models).