Use genuine IERS tables for Poisson series if possible.
There are several different formats for Poisson series tables in IERS
conventions. Older ones use several columns for higher degrees or for
putting different values in the same table (for example Δψ and Δε),
whereas newest ones use one table for each parameter and separate
degrees in different table sections introduced by some sub-headers like
j = 0  Nb of terms = 1306, j = 1  Nb of terms = 253 ...

When we introduced support for older IERS conventions a few weeks ago,
we backported the newer format to the older tables, meaning we had to
change the IERS files. This was cumbersome and error-prone. It also
appeared the older format was still use in recent conventions, like for
example table 5.1a in IERS conventions 2010. So this change adds support
for both Poisson series tables formats, and the embedded IERS files that
were in this format have been reverted to their original format.