First usable version of solid tides force model.
Reading and re-reading the IERS conventions, it seems that what we have
now in Orekit is compliant with the reference. However, some external
data seem to use an opposite sign on the sine part for the frequency
dependent part of ∆C₂₀. So we are not 100% confident to be correct and
have left a comment in the code at table loading points in the various
IERSConventions enumerate instances.

Changing the sign has a quite small effect: about 40cm after a one month
propagation on Low Earth Orbit. So it will probably not be immediate to
fully validate the correct sign.