Fixed solid Earth pole tide for IERS conventions 2010.
The mean pole model from table 7.7 used wrong units (arcseconds instead
of milli arcseconds).

An extensive comment has also been added concerning the sign discrepency
in the models between IERS 2003 and IERS 2010. In conventions 2003, the
equations read:

 ∆C₂₁ = −1.333 × 10⁻⁹ (m₁ − 0.0115m₂)
 ∆S₂₁ = −1.333 × 10⁻⁹ (m₂ + 0.0115m₁)

In conventions 2010, the equations read:

 ∆C₂₁ = −1.333 × 10⁻⁹ (m₁ + 0.0115m₂)
 ∆S₂₁ = −1.333 × 10⁻⁹ (m₂ − 0.0115m₁)

Some tests on a Low Earth Orbit have shown that the effect of this sign
change is about 0.4m after 30 days propagation, so it is quite small.
The proper sign is probably the one from conventions 2010, as it seems
consistent with the premises.