Added a DetectorEventHandler interface.
This interface is aimed at handling only the event occurrence part in
propagation. This allows:

 - to separate the event detection itself (which is declared by the
   EventDetector interface) from the action to perform once the event
   has been detected,
 - to avoid subclassing of events, which was cumbersome,
 - to share a single handler for several events.

The previous behavior with eventOccurred and resetState declared at
detector level and subclassing is still available but is deprecated and
will be removed in the next major release.

Thanks to Hank Grabowski for this new feature.