Read gravity files missing degree 0, or 1
Since the C and S coefficients for degree 0 and 1 are known, they are
initialized to their known values, if they are not specified in the gravity
file. Previously a missing coefficients exception was thrown if degree 0 or 1
was missing.

Some tests of the missing coefficient logic were updated since the only missing
coefficients in the test files were degree 0, or 1.