Fix set parameters in NumericalPropagatorBuilder
Due to how parameters were set in
NumericalPropagatorBuilder.buildPropagator(...) all parameters are set to the
last value. This error was not detected by the test cases because the test case
used Cd with an orbit that is not sensitive to drag.

Also fixed for TLEPropagatorBuilder, though it will have no effect since the
builder only supports one parameter.