Better implementation of projectToGround for ellipsoid.
The acceleration computed is more accurate, as the curvature of
the ellipsoid is properly computed in the sliding point direction.
So the component of the acceleration along the zenith axis induced
by horizontal velocity along a curved surface line is now correct.

There is still a problem for horizontal acceleration, though.