Added a protection against NaN appearing in EclipseDetector.
NaN may appear if the spacecraft enters inside either the occulting or
occulted body sphere. Of course, this should not appear with resonable
settings, but it occurred at least once in tests with an EventShifter
configured to used a huge unrealistic offset, for which the underlying
shiftedBy method was clearly not sufficient to meet accuracy needs.