Fixed target point in BodyCenterPointing attitude.
The target needs to be on body surface as the class is a specialization
of GroundPointing and could be used in YawCompensation mode. The target
point is now properly on surface, even if the line from spacecraft to
target also contains the body center, which is not the target.

Fixes issue #100.