Allowed slightly malformed TLE to be parsed.
We now accept TLE where the first tme derivative of mean motion is
shifted one character to the right (and has one less significat digit).
This does not match the description on celestrack, but it seems to
appear in the wild. The pattern accepted by isFormatOK has been adapted
for this special case (and only for this field), and more globally when
parsing doubles and integers, the start and end blanks are stripped
before replacement of ' ' with '0', thus avoid strings like '0+.1234"

Fixes issue #196