Circumvent strange behaviour of "mvn checkstyle:checkstyle".
Before the change, running "mvn clean" and then "mvn
checkstye:checkstyle" detected 4 checkstyle errors in
AbstractPropagatorConverter, at javadoc level, not being able to find
the class OrekitException! Running the same command after a "mvn
compile" (i.e. when all .class files are already in the maven target
directory) did not trigger any checkstyle error.

I'm not sure it is a real style problem or if it is a checkstyle bug.