Fixed a rare infinite loop in ellipsoid tessellation.
The loop occurred when some tiles created during one iteration of the
tessellation were so close to the zone of interest boundary that a very
thin stripes remained that was not empty, but the inside points were
closer to the boundary than the threshold. This sometimes lead to find a
start point around which a tile could be built, and an infinite loop as
nothing was done to cover this part, which remained indefinitely neither
empty not coverable.