Fix ephemeris start and end dates
Previously a generated ephemeris may not have been available at the start and
stop times used for propagation because of numerical precision issues. Now the
real start and end AbsoluteDates are used if they are equivalent to the double
representation of time. The public interfaces ModeHandler and StateMapper were
changed. Includes test cases for the numerical and DSST propagators.

Fixes #210