Removed a spurious test.
This test was not really meaningful, as it only checked that restarting
DSST did not consume too much time. It had been created after a bug was
identified a long time ago, but as it is only a timing issue, it relied
on monitoring time, which is inherently unstable. So the test did not
bring anything and it randomly failed, which was a nuisance as
continuous integration has been set up.