Improved DSST ephemeris consistency with propagator.
The propagator interpolates its short periodics elements using only
the current step, as it does not have information on the next step. So
the ephemeris should do the same, even if the next step is available
later on in the ephemeris.

In order to ensure this, we restrict the scope of short periodic terms
to a time slot.