Orbit determination now returns a configured propagator.
As users may need to propagate the estimated orbit, it is simpler for
users to directly get a fully configured propagator with the estimated
propagation parameters also set rather than get only the orbit. If only
the orbit is desired, it can be retrieved as the initial state of the
propagator.