Fixed wrong caching of past g function values.
The caching did "save" computation of g function values not only
when the same time was reused, but also for times one ulp away.
This broke events detection as when we are very close to a zero,
a one ulp difference in time changes the sign, by definition of
the crossing.