Add event tests from hipparchus, update algorithm
Ported a large suite of event tests from hipparchus ODE integrators to Orekit
Propagators. Created several sub tests to run these with different analytical
and numerical Propagators. Also included recent updates to to the Hipparchus
event finding algorithm.

At this time only the DP853 integrator seems to work reliably with events. The
Adams integrators have trouble with resets and the GBS integrator gets stuck in
an infinite loop. At this point I'm not sure if these are bugs per se or
limitations that need to be documented.