New interfaces for ephemeris files
Created new interfaces for reading and parsing ephemeris files in
EphemerisFile. The big new feature is the ability to create a propagator
directly from the information in the interface. The new interface also adds
support for ephemeris segments using that model that each ephemeris file
contains zero or more satellites and each satellite has zero or more ephemeris
segments. Also added some supporting classes for creating a bounded propagator
based on an ephemeris file, as well as some test cases.