Apply EphemerisFile interface to SP3 files
Changed SP3{Parser,File} to use the new interface and removed the OrbitFile
interface. Also added a test case to test the created propagator.

Changed the EphemerisFile interface so that SP3 coordinates can include clock
information. Also updated the EphemerisSegmentPropagator to work even if the
ephemeris is defined in a non-inertial frame.