Fixed shifting/interpolating transform providers with fields.
When using DerivativeSructure as field, the shifting and interpolating
transform providers rely on sampling on an underlying provider, plus
some dedicated interpolation formula. If the underlying provider already
computes the derivatives, then the interpolation formula should not add
the derivatives too, otherwise the derivatives would be computed (and
summed up) by several layers of code.

The error occurred in the previous version, and peaked with the GTOD
frame which uses both interpolation and shift: we end up having the
derivatives computed three times and the final value was the triple of
the correct value...

This change computes the derivatives (i.e. the Field specificities) at
the topmost level, i.e. in the interpolation formula. The sample itself
is based on raw Transform instances just wrapped to appear as
FieldTransform instances.