Fixed array index errors.
When compiling several Poisson series, not all series have all terms.
For example in the IERS conventions 2010 nutation series, the x and y
pole coordinates series have far more terms than the s+XY/2 series.
So some terms may not contribute to some coordinates, this was
considered for the value but was forgotten for the new (as of 9.0)
derivative.