Fix: honor CENTER_NAME in CCSDS ODM
In OEM files the getFrame() method ignored the value of CENTER_NAME which is
used in CCSDS files to change the center of the reference frame but keep the
rotational alignment. Added a method in ODMMetaData to compute the modified
frame if REF_FRAME is not already centered at CENTER_NAME. For this to work I
had to add a method to the CelestialBody interface for the new FieldTransform
methods. JPLCelestialBody already implemented the method. Also added a test case
and changed one that checked incorrect behavior.

Fix #334