Set up shift tests so they pass, but this is NOT satisfying!
Several tests appears to have too large errors (this is only
an intuition). I would have expected derivatives to improve
accuracy of the shiftedBy method, and in some case it does
not occur.

What is even more strange is that for CartesianOrbit, improvement
is real, so the data present in the derivatives is indeed enough
to shift the Cartesian orbits, so why doesn't it work to shift
Keplerian, circular or equinoctial orbits?

More work is needed on this issue, to either solve it and
get accurate shifts (which is what I expect), or to understand why
accuracy is not improved (which is what I fear).