Almost working version of simplified ForceModel API.
There is still something shady on radiation pressure force model.
Computing the lighting ratio with all associated derivatives leads to
tests failures, whereas ignoring the derivatives of this intermediate
variable works. This sounds illogical. We have to figure out what
happens there.

We also have to remove the former API with accelerationDerivative method
(which should be moved to the tests as it is used as a reference for the
new API).