Fixed derivatives computations near penumbra boundaries.
The RESET_DERIVATIVES action triggered by eclipses detectors at penumbra
entry/exit dates should happen within the penumbra zone, so derivatives
of the lighting ratio are correctly computed. These derivatives are
discontinuous because in full light or full umbra, the ratio is constant
and equald to either 1.0 or 0.0, so the derivatives are zero. In
penumbra however, the ratio is highly dependent on position as it goes
from 1.0 to 0.0 in just a few kilometers. The event detectors are now
triggered using a very small angular margin (1.0e-10 radians), so
derivatives computed at event triggering date are considered to be
inside the penumbra, and the derivatives are non-zero.