Exceptions fixed.
NumericalPropagator: Added if in to call a TimeStampedPVCoordinates
constructor without acceleration acceleration is unknown.

testCaseJupiter: removed three final Asserts that appeared at the end as
to make it a Tutorial.

testCastL2_DRAFT: removed OrbitType definition as Cartesian, null
instead as to follow absolute position-velocity-acceleration.

testCaseRotatingEarth: definition of orbit in case 2 changed to define it AbsolutePVCoordinates. Tolerances are built using referenceOrbit. OrbitType to null for the code follow the absolute position-velocity-acceleration path.