Compute as few derivatives as possible in orbit determination.
As some force models depend only on position (gravity, third body,
parametric acceleration when direction is inertial), we can reduce
the number of derivatives computed for these force models.
The other force models that depend on velocity (either directly or due
to a dependency on attitude which itself may depend on velocity,
for example for LOF offset attitudes) are computed with all derivatives
taken into account.