L2TransformProviderTest created and changes on L2TranformProvider
L2TransformProvider is now cleaner. No output frame, L2 point is always
given in primary body's inertially oriented frame. Also changes in style
and variable definition simplified. 

L2TransformProviderTest created to see if the L2 point and the secondary
body are aligned with the primary body. Still with mistakes that must be
solved.