L2Frame implemented and L2TransformProviderTest finished
Implemented class to create L2Frame using Frame constructor and imposing
primaryBody's inertially oriented frame as parent Frame. 


L2TransformProviderTest finished having tested orientation and distance
for Earth-Moon, Sun-Earth and Sun-Jupiter. This led to a .negate() to be
added to the getL2() method when creating the Transform in
L2TransformProvider.