Pre-sort main tides.
The mains tides used in Onsala Space Observatory files in BLQ format
are fixed and their sort order according to rate is stable over a very
large time range. Even considering all the 342 tides used for complete
model and not only the 11 main tides, sort order is stable for about
180000 years, from -122502-11-09 (cross-over between tides 245556 and
245635) to 56824-11-02 (cross-over between tides 274554 (R₂) and
274556).

So there is no need to attempt run-time sorting. We know the 11 main
tides in BLQ format will always be grouped as::
  - 3 long period tides (Ssa, Mm, Mf in increasing rate order),
  - 4 diurnal tides (Q₁, O₁, P₁, K₁ in increasing rate order),
  - 4 semi-diurnal tides (N₂, M₂, S₂, K₂ in increasing rate order).
We take this into account in the BLQ parser and provide the parsed
coefficients ready to be used for spline interpolation.