Inhibate reference fortran state preservation between calls.
The reference fortran program computes the eclipse times from the first
call within eclipse, and reuses this in subsequent calls. Orekit design
choice is to avoid stateful attitude modes, so we redo the computation
on each call, which is not  problem since the computation are very
lightweight. So we cleared the fortran arrays while generating the
reference data.