Allow GNSS attitude to cope with turns extending after eclipse span.
When the β angle is small, the turn can be vary large, up to 180°. For
satellite with small maximum rotation rate, such a large turn take a lot
of time and the recovery of nominal attitude at the end of the turn can
occur outside of the eclipse span. We now store identified turns time
span so we can detect if we are still in the recovery phase or already
back to nominal yaw steering mode.