Make OrekitException extend RuntimeException.
This change has been discussed on both the developers and users lists,
and results from a poll among users. The rationale of the change is
base on three observations:

 1) before the change OrekitException was a checked exception.
    Therefore, calling features that throws one from some higher
    level library may imply wrapping the checked exception. This
    typically occurs when optimizing some parts (maneuvers, orbit
    determination) with Hipparchus calling Orekit, or calling an
    Orekit function using Java 8 lambda functions.
 2) most if not all Orekit features do throw such an exception,
    so the information to the user that a specific function does
    throws an OrekitException was pointless as all of them do.
 3) in many cases it is impossible to recover from these exceptions
    (for example if UTC-TAI or JPL data is missing, one can barely do
    anything in Orekit). Therefore, basically all exceptions thrown
    from Orekit classes bubbled up to the top level application and
    were caught there just to be displayed at the end.

Observations 2 and 3 lead to think that unchecked exceptions are better
suited than unchecked exceptions for Orekit expected use cases, and
observation 1 lead to think that unchecked exceptions would greatly
simplify user code.

So we moved from checked exceptions to unchecked exceptions.

Fixes #484.