Replaced shiftedBy by proper propagation in reference data generator.
Using shiftedBy led the β angle to be wrong for use cases with large
time span (like Beidou-2I and Beidou-2M). Using proper propagation
required to generate data in two passes, as we needed to aggregate SP3
files when time span crosses the midnight boundary.

With this change, the Orekit implementation of GNSS attitude laws is
considered to be fully validated!