Added scheduling predicates that can be combined.
The predicates by themselves are simple but can be combined to
match mission-specific or measurement-specific needs. As an
example an optical tracking of satellites on the stars background
using a telescope requires to have at measurement time the satellite
visible from the telescope location, the telescope location be
in the night side of the Earth and the satellite being still sunlit.
This can be described as a combing scheduling predicate that is the
AND combination of VisibleFromGround, GroundAtNight and Sunlit
basic predicates.