Improved javadoc.
The previous explanation for multi-threading and propagation set up for
fixing #479 was incomplete. The fact is that propagators cannot be
called from several threads at once (they do preserve some internal
state), but they CAN be used in multi-threaded applications as long as
one propagator is only called from one thread. We even have a helper
class for that: PropagatorsParallelizer.

Fixes #479