Change itrf-versions.conf to match based on prefix
Previous method of matching based on a regular expression could lead to an
application hang if an crafted file was loaded. Since these files can be loaded
from the network the matching mechanism was changed to avoid a potential denial
of service. Also updated the separator in ZipJarCrawler to match the convention
in JarURLConnection. Added some test cases.

Fixes #543