Added on the fly shrinking of search ellipsoid on LAMBDA method.
This *vastly* improves search time. As an example, the test for random
problems was pushed up to dimension 20 (previously the test was limited
to dimension 10 so it completes in reasonable time). Without ellipsoid
shrinking, it involved hundreds of millions of candidate points for some
problems and I interrupted it after 45 minutes. Now it involves at most
20 candidate points for all problems and the test completes in 3.6
seconds...