Replace sleep with timeout
Using sleep imposes to wait the given amount of time
even if the result is available sooner.
Using the TimeOut, the get is released as soon as
the result is available. Thus, we can use a greater value
for the timeout to handle the case the execution
is exceptionnaly long, without any impact on faster conditions.