Update AbsoluteDate for TimeScales
Updated AbsoluteDate to provide alternatives to using the default
DataContext, including the constant epochs. Also updated
FieldAbsoluteDate and all classes that used the methods dependent on the
default DataContext. Added AbsoluteDate.ARBITRARY_EPOCH to replace
places J2000_EPOCH was used but the point in time was irrelevant. Added
DateTimeComponents.JULIAN_EPOCH to use it separated from the default TT
time scale.

Updated the TLEConverterTest to use a looser tolerance. The changes
resulted in a 1 ulp difference in one of the calculations in
SDP4/DeepSDP4. The test had been using a tolerance less than its
convergence threshold which makes the pass/fail fairly arbitrary.

Part of #607