Fix adding EOP loaders before leap seconds
In Orekit 10.0 it is possible to add default and custom EOP loaders
before leap seconds are available to the application. Before this commit
UTC was a constructor argument to the loaders so it was needed before
the loaders were added. Restored the 10.0 behavior by changing the
constructor argument to a UTC supplier that is only accessed when it is
first needed.

Part of #607