Fix missing doc of default data context use
The toString() methods of CartesianOrbit, FieldCartesianOrbit,
TimeStampedFieldPVCoordinates, and TimeStampedPVCoordinates used the
default data context by appending a date to a string. Modified
CartesianOrbit to only use the six defining elements like is done for
the other subclasses of Orbit. Added a toString(TimeScale) method to
TimeStampedPVCoordinates and modified the field version to call it.

Part of #607