Update existing tests
The purpose of testBug74() is to test 3-digit orbit parameters. However, because of the new range check, the test now fails. 
Solution: reducing the parameters to the acceptable range, but still 3-digit.

testBug77TooLargeEccentricity() also caused issue because of the new check on the eccentricity which raises a different exception as previously expected, now MathIllegalArgumentException