Update existing tests
The purpose of testBug74() is to test 3-digit orbit parameters. However, because of the new range check, the test now fails. 
Solution: reducing the parameters to the acceptable range, but still 3-digit.

testBug77TooLargeEccentricity() also caused an issue because the check on the eccentricity is now done in the constructor and raises an OrekitException with a different error message than before